feat(web): move navy header into shared layout so all pages match commongood.earth
